build, test & README
The purpose of this commit is to clean up the work done up to this point,
add proper unit testing, and create a single "command centre".
The role of the "command centre" is the shell script "project", which implements
such options as "--test", or "--derive", with the option "--generateGraph" being still a todo.