test sim creation, fix imports, other improvements
It took a lot of work and some hackery to fix the imports, so that both
1) tests can be executed outside of python, i.e. using -m flag (but not without).
2) module and submodules can be imported as a standard python module.
3) things are in separate files for easier maintanance.
4) but when the parent module is imported everything lives in one namespace.

The solution is to do some __init__.py hackery, create a separate submodule for tests and create an empty __main__.py inside that submodule.

Writing test cases for sims revealed a bug in moveOnSphere from pointPicking.py, some octants were getting a wrong sine. It turned out that the solution was to use atan2, and simplified verision of the formula from the website http://www.movable-type.co.uk/scripts/latlong.html. The new formula still passes old test cases, so that's good.

a function was added to turn lam, psi -> x, y, z. It was useful for testing.