Finished a set of functions to generate graphs, and the first graph is generated by one of the test cases.
The next step is to implement graph processing.

Fixed sim breeding, and its test case, which were both broken due to
the assumption that sims had a parameter "gender", when in fact they have
"isMale".

Added a function to serialise sims to tuples.

While working on population graph-to-file serialisation I had to work out a couple of non-trivial issues revolving the following constraints:
    1. Json couldn't be used, because tuple is not supported in json
    2. I had to use tuple because of networkx, which requires each node to be hashable.
    3. Subsequently I've noticed that python's tuple of tuples has an interesting feature, that, roughly speaking, tuplesOfTuples == eval(str(tupleOfTuples)).
    4. Based on this I decided to eval tuples from a file, line by line.
    5. I realised that serialising a graph from memory would be difficult, because caution is required to not include nodes multiple times (for example breadth first would be tricky)
    6. It was then decided that each graph can be converted into a canoncial (lexicographically sorted) tuple representation, which can be saved into a file as per #3

More work was carried out on executing test cases deterministically. It turned out that executing python module isn't a good idea overall, so running test cases is now the responsibility of test.py, which can be used to run test cases deterministically.